+++ src/components/checkbox/checkbox.component.spec.ts
@@ -401,6 +401,24 @@ describe('Checkbox', () => {
             )
         );
 
+        it('can be disabled via the form control',
+            componentTest(() => TestComponent, `
+                <form [formGroup]="testForm">
+                    <gtx-checkbox formControlName="testControl"></gtx-checkbox>
+                </form>`,
+                (fixture, instance) => {
+                    fixture.detectChanges();
+
+                    const input: HTMLInputElement = fixture.nativeElement.querySelector('input');
+                    expect(input.disabled).toBe(false);
+
+                    instance.testForm.get('testControl').disable();
+                    fixture.detectChanges();
+                    expect(input.disabled).toBe(true);
+                }
+            )
+        );
+
     });
 });