+++ src/components/range/range.component.spec.ts
@@ -210,6 +210,24 @@ describe('Range:', () => {
             )
         );
 
+        it('can be disabled via the form control',
+            componentTest(() => TestComponent, `
+                <form [formGroup]="testForm">
+                    <gtx-range formControlName="test"></gtx-range>
+                </form>`,
+                (fixture, instance) => {
+                    fixture.detectChanges();
+
+                    const input: HTMLInputElement = fixture.nativeElement.querySelector('input');
+                    expect(input.disabled).toBe(false);
+
+                    instance.testForm.get('test').disable();
+                    fixture.detectChanges();
+                    expect(input.disabled).toBe(true);
+                }
+            )
+        );
+
     });
 
     describe('DOM Events:', () => {